+++ packages/wx-eventemitter/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/areslabs/alita/compare/v2.4.6...v2.5.0) (2020-04-20)
+
+**Note:** Version bump only for package @areslabs/wx-eventemitter
+
+
+
+
+
+## [2.4.6](https://github.com/areslabs/alita/compare/v2.4.5...v2.4.6) (2020-04-16)
+
+**Note:** Version bump only for package @areslabs/wx-eventemitter
+
+
+
+
+
+## [2.4.5](https://github.com/areslabs/alita/compare/v2.4.4...v2.4.5) (2020-04-10)
+
+**Note:** Version bump only for package @areslabs/wx-eventemitter
+
+
+
+
+
+## [2.4.4](https://github.com/areslabs/alita/compare/v2.4.3...v2.4.4) (2020-04-10)
+
+**Note:** Version bump only for package @areslabs/wx-eventemitter
+
+
+
+
+
+## [2.4.3](https://github.com/areslabs/alita/compare/v2.4.2...v2.4.3) (2020-04-02)
+
+**Note:** Version bump only for package @areslabs/wx-eventemitter
+
+
+
+
+
 ## [2.4.2](https://github.com/areslabs/alita/compare/v2.4.1...v2.4.2) (2020-03-19)
 
 **Note:** Version bump only for package @areslabs/wx-eventemitter