+++ packages/wx-react/CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.5.0](https://github.com/areslabs/alita/compare/v2.4.6...v2.5.0) (2020-04-20)
+
+
+### Features
+
+* **alita-core wx-react:** 支持 onLayout ([6f9cd2d](https://github.com/areslabs/alita/commit/6f9cd2df8abbca74212236a6ebb891e5c77a37fa))
+
+
+
+
+
+## [2.4.6](https://github.com/areslabs/alita/compare/v2.4.5...v2.4.6) (2020-04-16)
+
+**Note:** Version bump only for package @areslabs/wx-react
+
+
+
+
+
+## [2.4.5](https://github.com/areslabs/alita/compare/v2.4.4...v2.4.5) (2020-04-10)
+
+**Note:** Version bump only for package @areslabs/wx-react
+
+
+
+
+
+## [2.4.4](https://github.com/areslabs/alita/compare/v2.4.3...v2.4.4) (2020-04-10)
+
+**Note:** Version bump only for package @areslabs/wx-react
+
+
+
+
+
+## [2.4.3](https://github.com/areslabs/alita/compare/v2.4.2...v2.4.3) (2020-04-02)
+
+
+### Bug Fixes
+
+* **wx-react  example:** 紧急修复 metro-react-native-babel-preset升级造成的小程序运行bug ([67b635c](https://github.com/areslabs/alita/commit/67b635cebebd03e2e4745d1f0f4aeabf09ecaf0f))
+
+
+
+
+
 ## [2.4.2](https://github.com/areslabs/alita/compare/v2.4.1...v2.4.2) (2020-03-19)
 
 **Note:** Version bump only for package @areslabs/wx-react